+++ exsampleFiles/Debug.py
@@ -2,7 +2,7 @@ from GabesPythonToolBox.Utility.Debug import *
 from GabesPythonToolBox.Utility.Logger import *
 
 # Enable debugging
-Debug.set_debug_enabled(True,verbosity=4)
+Debug.set_debug_enabled(True,verbosity=1)
 
 # Add some custom message groups
 Debug.add_group('LIB', True)
@@ -16,6 +16,7 @@ Debug.log("------------------------------------------", message_type="Info", gro
 Debug.log("Custom Header Message", message_type="Header")
 Debug.log("This is an error message", message_type="Error")
 Debug.log("This is a failed message", message_type="Fail")
+Debug.log("This is a Unachievable message", message_type="Unachievable")
 Debug.log("This is a success message", message_type="Success")
 Debug.log("This is a warning message", message_type="Warning")
 Debug.log("This is an info message", message_type="Info")
@@ -25,6 +26,8 @@ Debug.log("This is an End message", message_type="End")
 Debug.log("This has a wrong type", message_type="-")
 
 Debug.log("------------------------------------------", message_type="Info", group="Showcase")
+Debug.log("This is a New style message", message_type=LogType.Success)
+Debug.log("This is a Old style message", message_type="Success")
 Debug.log("\n\n\n", message_type="None", group="Showcase")
 
 # Using a group filter
@@ -37,6 +40,10 @@ Debug.log("This should now be shown since 'Database' is enabled", message_type="
 
 Debug.disable_group('LIB')
 Debug.log("This should NOT be shown since 'LIB' is now disabled", message_type="Info", group="LIB")
+
+Debug.enable_group('LIB')
+Debug.log("This should be shown since 'LIB' is now enabled", message_type="Info", group="LIB")
+
 
 #list group
 Debug.log("list groups", message_type="Info", group="Showcase")
@@ -55,4 +62,5 @@ Debug.remove_group("test")
 
 Debug.log("list groups", message_type="Info", group="Showcase")
 Debug.log(Debug.list_groups(), message_type="Info", group="Showcase")
+